style: refactor package structure for business logic clasees
