+++ src/main/java/calculator/view/InputView.java
@@ -3,6 +3,6 @@ package calculator.view;
 public class InputView {
 
     public void showStartMessage() {
-        System.out.println("덧셈할 문자열을 입력해주세요.\n기본 구분자는 쉼표(,)와 콜론(:)입니다.");
+        System.out.println("덧셈할 문자열을 입력해주세요.\n계산은 0이상의 정수만 가능합니다.\n기본 구분자는 쉼표(,)와 콜론(:)입니다.");
     }
 }
